freelance_platform/.venv-requirements.txt
-r requirements.txt
pylint>=3.0.0
pylint-django>=2.5.0
pytest>=7.0.0
pytest-django>=4.5.0
black>=23.0.0
isort>=5.10.0
flake8>=5.0.0
coverage>=6.0.0
django-debug-toolbar>=4.0.0 